% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{ggtheme}
\alias{ggtheme}
\title{Custom ggplot Theme}
\usage{
ggtheme(base_size = 8, legend_position = "default")
}
\arguments{
\item{base_size}{Number specifying base font size}

\item{legend_position}{Character specifying legend position (\code{"topleft"} or \code{"topright"} or \code{"lowleft"} or \code{"lowright"} or \code{"topmid"} or \code{"lowmid"})}
}
\value{
Returns a ggplot theme
}
\description{
Modifies theme of ggplot object. White background with grey major grid lines and easily moveable legend.
}
