% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{package_set}
\alias{package_set}
\title{Package sets}
\usage{
package_set(pkgs = "all")
}
\arguments{
\item{package_set}{Character string specifying set of packages to load (\code{"spatial"} or \code{"stats"} or \code{"all"})}
}
\value{
Loads packages, fixes some obvious name clashes
}
\description{
Loads some useful package sets, fixes some obvious name clashes
}
\details{
Spatial packages are \code{raster}, \code{sp}, \code{rgdal}, \code{maptools}, \code{mapproj}, \code{geosphere}, \code{rgeos} and \code{spdplyr}.
The \code{raster:::select} function is overwritten with \code{dplyr:::select}.\cr\cr
Stats packages are \code{lme4} and \code{MuMIn}
}
